ctabganplus/requirements.txt
numpy==1.26.3
python==3.10.13
ctgan==0.9.1
joblib==1.2.0
matplotlib==3.8.3
pillow==10.2.0
nbformat==5.10.4
pip==23.3.1
plotly==5.20.0
torch==1.9.1
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.4
sdmetrics==0.13.0
dython==0.7.3
scipy==1.4.1
seaborn==0.13.0
sdv==0.13.2
table-evaluator==1.6.1
kaleido==0.1.0post1
torchinfo
torch==2.2.1+cu121
wandb=0.16.6
tqdm==4.66.2
